# Add option to resize with out setting nulls