--- a/velox/vector/ConstantVector.h
+++ b/velox/vector/ConstantVector.h
@@ -283,7 +283,7 @@ class ConstantVector final : public SimpleVector<T> {
     return index_;
   }
 
-  void resize(vector_size_t size) override {
+  void resize(vector_size_t size, bool /*unused*/) override {
     BaseVector::length_ = size;
   }