--- a/velox/vector/SimpleVector.h
+++ b/velox/vector/SimpleVector.h
@@ -215,7 +215,7 @@ class SimpleVector : public BaseVector {
     return max_;
   }
 
-  void resize(vector_size_t size) override {
+  void resize(vector_size_t size, bool setNotNull = true) override {
     VELOX_CHECK(false, "Can only resize flat vectors.");
   }