… velox/expression/tests/ArrayProxyTest.cpp