@mbasmanova 2022-01-06
@laithsakka Laith, would you extract this change into a separate PR so it can land sooner? It would be nice to document the new 'setNotNull' parameter.